+++ e2e/coverage-provider-v8/with-resetModules/uncovered.js
@@ -5,5 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface obj {}
+module.exports = {};