+++ packages/jest-circus/src/run.ts
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import throat from 'throat';
+import pLimit = require('p-limit');
 import type {Circus} from '@jest/types';
 import {dispatch, getState} from './state';
 import {RETRY_TIMES} from './types';
@@ -46,12 +46,13 @@ const _runTestsForDescribeBlock = async (
 
   if (isRootBlock) {
     const concurrentTests = collectConcurrentTests(describeBlock);
-    const mutex = throat(getState().maxConcurrency);
+    const mutex = pLimit(getState().maxConcurrency);
     for (const test of concurrentTests) {
       try {
         const promise = mutex(test.fn);
         // Avoid triggering the uncaught promise rejection handler in case the
         // test errors before being awaited on.
+        // eslint-disable-next-line @typescript-eslint/no-empty-function
         promise.catch(() => {});
         test.fn = () => promise;
       } catch (err) {