+++ packages/jest-mock/src/index.ts
@@ -160,6 +160,7 @@ export interface MockInstance<T extends FunctionLike = UnknownFunction> {
   mockRejectedValueOnce(value: RejectType<T>): this;
 }
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface SpyInstance<T extends FunctionLike = UnknownFunction>
   extends MockInstance<T> {}
 
@@ -230,7 +231,7 @@ type MockFunctionConfig = {
 
 const MOCK_CONSTRUCTOR_NAME = 'mockConstructor';
 
-const FUNCTION_NAME_RESERVED_PATTERN = /[\s!-\/:-@\[-`{-~]/;
+const FUNCTION_NAME_RESERVED_PATTERN = /[\s!-/:-@[-`{-~]/;
 const FUNCTION_NAME_RESERVED_REPLACE = new RegExp(
   FUNCTION_NAME_RESERVED_PATTERN.source,
   'g',
@@ -633,6 +634,7 @@ export class ModuleMocker {
           metadata.members.prototype.members) ||
         {};
       const prototypeSlots = this._getSlots(prototype);
+      // eslint-disable-next-line @typescript-eslint/no-this-alias
       const mocker = this;
       const mockConstructor = matchArity(function (
         this: ReturnType<T>,
@@ -896,7 +898,7 @@ export class ModuleMocker {
   ): Mock<T> {
     // metadata not compatible but it's the same type, maybe problem with
     // overloading of _makeComponent and not _generateMock?
-    // @ts-expect-error
+    // @ts-expect-error - unsure why TSC complains here?
     const mock = this._makeComponent(metadata);
     if (metadata.refID != null) {
       refs[metadata.refID] = mock;
@@ -1020,6 +1022,7 @@ export class ModuleMocker {
   isMockFunction<T extends FunctionLike = UnknownFunction>(
     fn: SpyInstance<T>,
   ): fn is SpyInstance<T>;
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-constraint
   isMockFunction<P extends Array<unknown>, R extends unknown>(
     fn: (...args: P) => R,
   ): fn is Mock<(...args: P) => R>;
@@ -1087,7 +1090,9 @@ export class ModuleMocker {
     if (!this.isMockFunction(original)) {
       if (typeof original !== 'function') {
         throw new Error(
-          `Cannot spy the ${methodName} property because it is not a function; ${this._typeOf(
+          `Cannot spy the ${String(
+            methodName,
+          )} property because it is not a function; ${this._typeOf(
             original,
           )} given instead`,
         );
@@ -1149,7 +1154,9 @@ export class ModuleMocker {
 
     if (!obj) {
       throw new Error(
-        `spyOn could not find an object to spy upon for ${propertyName}`,
+        `spyOn could not find an object to spy upon for ${String(
+          propertyName,
+        )}`,
       );
     }
 
@@ -1166,16 +1173,18 @@ export class ModuleMocker {
     }
 
     if (!descriptor) {
-      throw new Error(`${propertyName} property does not exist`);
+      throw new Error(`${String(propertyName)} property does not exist`);
     }
 
     if (!descriptor.configurable) {
-      throw new Error(`${propertyName} is not declared configurable`);
+      throw new Error(`${String(propertyName)} is not declared configurable`);
     }
 
     if (!descriptor[accessType]) {
       throw new Error(
-        `Property ${propertyName} does not have access type ${accessType}`,
+        `Property ${String(
+          propertyName,
+        )} does not have access type ${accessType}`,
       );
     }
 
@@ -1184,7 +1193,9 @@ export class ModuleMocker {
     if (!this.isMockFunction(original)) {
       if (typeof original !== 'function') {
         throw new Error(
-          `Cannot spy the ${propertyName} property because it is not a function; ${this._typeOf(
+          `Cannot spy the ${String(
+            propertyName,
+          )} property because it is not a function; ${this._typeOf(
             original,
           )} given instead`,
         );
@@ -1199,7 +1210,7 @@ export class ModuleMocker {
       (descriptor[accessType] as Mock<() => T>).mockImplementation(function (
         this: unknown,
       ) {
-        // @ts-expect-error
+        // @ts-expect-error - wrong context
         return original.apply(this, arguments);
       });
     }